+++ lib/src/easy_ads.dart
@@ -1,17 +1,22 @@
 import 'dart:async';
 
+import 'package:applovin_max/applovin_max.dart';
 import 'package:collection/collection.dart';
 import 'package:easy_ads_flutter/easy_ads_flutter.dart';
 import 'package:easy_ads_flutter/src/easy_admob/easy_admob_interstitial_ad.dart';
 import 'package:easy_ads_flutter/src/easy_admob/easy_admob_rewarded_ad.dart';
-import 'package:easy_ads_flutter/src/easy_applovin/easy_applovin_ad.dart';
+import 'package:easy_ads_flutter/src/easy_applovin/easy_applovin_banner_ad.dart';
+import 'package:easy_ads_flutter/src/easy_applovin/easy_applovin_interstitial_ad.dart';
+import 'package:easy_ads_flutter/src/easy_applovin/easy_applovin_rewarded_ad.dart';
 import 'package:easy_ads_flutter/src/easy_facebook/easy_facebook_banner_ad.dart';
 import 'package:easy_ads_flutter/src/easy_facebook/easy_facebook_full_screen_ad.dart';
 import 'package:easy_ads_flutter/src/easy_unity/easy_unity_ad.dart';
+import 'package:easy_ads_flutter/src/utils/auto_hiding_loader_dialog.dart';
 import 'package:easy_ads_flutter/src/utils/easy_event_controller.dart';
 import 'package:easy_ads_flutter/src/utils/easy_logger.dart';
 import 'package:easy_ads_flutter/src/utils/extensions.dart';
-import 'package:facebook_audience_network/facebook_audience_network.dart';
+import 'package:easy_audience_network/easy_audience_network.dart';
+import 'package:flutter/material.dart';
 import 'package:unity_ads_plugin/unity_ads_plugin.dart';
 
 class EasyAds {
@@ -21,11 +26,15 @@ class EasyAds {
   /// Google admob's ad request
   AdRequest _adRequest = const AdRequest();
   late final IAdIdManager adIdManager;
+  late AppLifecycleReactor _appLifecycleReactor;
 
   final _eventController = EasyEventController();
   Stream<AdEvent> get onEvent => _eventController.onEvent;
 
   List<EasyAdBase> get _allAds => [..._interstitialAds, ..._rewardedAds];
+
+  /// All the interstitial ads will be stored in it
+  final List<EasyAdBase> _appOpenAds = [];
 
   /// All the interstitial ads will be stored in it
   final List<EasyAdBase> _interstitialAds = [];
@@ -36,6 +45,10 @@ class EasyAds {
   /// [_logger] is used to show Ad logs in the console
   final EasyLogger _logger = EasyLogger();
 
+  /// On banner, ad badge will appear
+  bool get showAdBadge => _showAdBadge;
+  bool _showAdBadge = false;
+
   /// Initializes the Google Mobile Ads SDK.
   ///
   /// Call this method as early as possible after the app launches
@@ -44,12 +57,18 @@ class EasyAds {
   Future<void> initialize(
     IAdIdManager manager, {
     bool unityTestMode = false,
+    bool fbTestMode = false,
+    bool isShowAppOpenOnAppStateChange = false,
     AdRequest? adMobAdRequest,
     RequestConfiguration? admobConfiguration,
     bool enableLogger = true,
     String? fbTestingId,
+    bool isAgeRestrictedUserForApplovin = false,
     bool fbiOSAdvertiserTrackingEnabled = false,
+    int appOpenAdOrientation = AppOpenAd.orientationPortrait,
+    bool showAdBadge = false,
   }) async {
+    _showAdBadge = showAdBadge;
     if (enableLogger) _logger.enable(enableLogger);
     adIdManager = manager;
     if (adMobAdRequest != null) {
@@ -60,31 +79,42 @@ class EasyAds {
       MobileAds.instance.updateRequestConfiguration(admobConfiguration);
     }
 
-    if (manager.fbAdIds?.appId != null) {
+    final fbAdId = manager.fbAdIds?.appId;
+    if (fbAdId != null && fbAdId.isNotEmpty) {
       _initFacebook(
         testingId: fbTestingId,
+        testMode: fbTestMode,
         iOSAdvertiserTrackingEnabled: fbiOSAdvertiserTrackingEnabled,
         interstitialPlacementId: manager.fbAdIds?.interstitialId,
         rewardedPlacementId: manager.fbAdIds?.rewardedId,
       );
     }
 
-    if (manager.admobAdIds?.appId != null) {
+    final admobAdId = manager.admobAdIds?.appId;
+    if (admobAdId != null && admobAdId.isNotEmpty) {
       final response = await MobileAds.instance.initialize();
       final status = response.adapterStatuses.values.firstOrNull?.state;
+
+      response.adapterStatuses.forEach((key, value) {
+        _logger.logInfo(
+            'Google-mobile-ads Adapter status for $key: ${value.description}');
+      });
 
       _eventController.fireNetworkInitializedEvent(
           AdNetwork.admob, status == AdapterInitializationState.ready);
 
       // Initializing admob Ads
-      EasyAds.instance._initAdmob(
+      await EasyAds.instance._initAdmob(
+        appOpenAdUnitId: manager.admobAdIds?.appOpenId,
         interstitialAdUnitId: manager.admobAdIds?.interstitialId,
         rewardedAdUnitId: manager.admobAdIds?.rewardedId,
+        appOpenAdOrientation: appOpenAdOrientation,
+        isShowAppOpenOnAppStateChange: isShowAppOpenOnAppStateChange,
       );
     }
 
     final unityGameId = manager.unityAdIds?.appId;
-    if (unityGameId != null) {
+    if (unityGameId != null && unityGameId.isNotEmpty) {
       EasyAds.instance._initUnity(
         unityGameId: unityGameId,
         testMode: unityTestMode,
@@ -94,8 +124,11 @@ class EasyAds {
     }
 
     final appLovinSdkId = manager.appLovinAdIds?.appId;
-    if (appLovinSdkId != null) {
+    if (appLovinSdkId != null && appLovinSdkId.isNotEmpty) {
       EasyAds.instance._initAppLovin(
+        sdkKey: appLovinSdkId,
+        keywords: adMobAdRequest?.keywords,
+        isAgeRestrictedUser: isAgeRestrictedUserForApplovin,
         interstitialAdUnitId: manager.appLovinAdIds?.interstitialId,
         rewardedAdUnitId: manager.appLovinAdIds?.rewardedId,
       );
@@ -108,13 +141,8 @@ class EasyAds {
   /// [adSize] is used to provide ad banner size
   EasyAdBase? createBanner(
       {required AdNetwork adNetwork, AdSize adSize = AdSize.banner}) {
-    assert(
-        adNetwork == AdNetwork.unity ||
-            adNetwork == AdNetwork.admob ||
-            adNetwork == AdNetwork.facebook,
-        'Only admob, unity and facebook banners are available right now');
-
     EasyAdBase? ad;
+
     switch (adNetwork) {
       case AdNetwork.admob:
         final bannerId = adIdManager.admobAdIds?.bannerId;
@@ -136,11 +164,21 @@ class EasyAds {
         }
         break;
       case AdNetwork.facebook:
-        final bannerId = adIdManager.fbAdIds?.interstitialId;
+        final bannerId = adIdManager.fbAdIds?.bannerId;
         assert(bannerId != null,
             'You are trying to create a banner and Facebook Banner id is null in ad id manager');
         if (bannerId != null) {
-          return EasyFacebookBannerAd(bannerId, adSize: adSize);
+          ad = EasyFacebookBannerAd(bannerId, adSize: adSize);
+          _eventController.setupEvents(ad);
+        }
+        break;
+      case AdNetwork.appLovin:
+        final bannerId = adIdManager.appLovinAdIds?.bannerId;
+        assert(bannerId != null,
+            'You are trying to create a banner and Applovin Banner id is null in ad id manager');
+        if (bannerId != null) {
+          ad = EasyApplovinBannerAd(bannerId);
+          _eventController.setupEvents(ad);
         }
         break;
       default:
@@ -150,9 +188,12 @@ class EasyAds {
   }
 
   Future<void> _initAdmob({
+    String? appOpenAdUnitId,
     String? interstitialAdUnitId,
     String? rewardedAdUnitId,
     bool immersiveModeEnabled = true,
+    bool isShowAppOpenOnAppStateChange = true,
+    int appOpenAdOrientation = AppOpenAd.orientationPortrait,
   }) async {
     // init interstitial ads
     if (interstitialAdUnitId != null &&
@@ -176,34 +217,64 @@ class EasyAds {
 
       await ad.load();
     }
+
+    if (appOpenAdUnitId != null &&
+        _appOpenAds.doesNotContain(AdNetwork.admob, AdUnitType.appOpen)) {
+      final appOpenAdManager =
+          EasyAdmobAppOpenAd(appOpenAdUnitId, _adRequest, appOpenAdOrientation);
+      await appOpenAdManager.load();
+      if (isShowAppOpenOnAppStateChange) {
+        _appLifecycleReactor =
+            AppLifecycleReactor(appOpenAdManager: appOpenAdManager);
+        _appLifecycleReactor.listenToAppStateChanges();
+      }
+      _appOpenAds.add(appOpenAdManager);
+      _eventController.setupEvents(appOpenAdManager);
+    }
   }
 
   Future<void> _initAppLovin({
+    required String sdkKey,
+    bool? isAgeRestrictedUser,
     String? interstitialAdUnitId,
     String? rewardedAdUnitId,
+    List<String>? keywords,
   }) async {
+    final response = await AppLovinMAX.initialize(sdkKey);
+
+    AppLovinMAX.targetingData.maximumAdContentRating =
+        isAgeRestrictedUser == true
+            ? AdContentRating.allAudiences
+            : AdContentRating.none;
+
+    if (keywords != null) {
+      AppLovinMAX.targetingData.keywords = keywords;
+    }
+
+    if (response != null) {
+      _eventController.fireNetworkInitializedEvent(AdNetwork.appLovin, true);
+    } else {
+      _eventController.fireNetworkInitializedEvent(AdNetwork.appLovin, false);
+    }
+
     // init interstitial ads
     if (interstitialAdUnitId != null &&
         _interstitialAds.doesNotContain(
             AdNetwork.appLovin, AdUnitType.interstitial)) {
-      final ad = EasyApplovinFullScreenAd(
-          interstitialAdUnitId, AdUnitType.interstitial);
+      final ad = EasyApplovinInterstitialAd(interstitialAdUnitId);
       _interstitialAds.add(ad);
       _eventController.setupEvents(ad);
 
-      await ad.initialize();
       await ad.load();
     }
 
     // init rewarded ads
     if (rewardedAdUnitId != null &&
         _rewardedAds.doesNotContain(AdNetwork.appLovin, AdUnitType.rewarded)) {
-      final ad =
-          EasyApplovinFullScreenAd(rewardedAdUnitId, AdUnitType.rewarded);
+      final ad = EasyApplovinRewardedAd(rewardedAdUnitId);
       _rewardedAds.add(ad);
       _eventController.setupEvents(ad);
 
-      await ad.initialize();
       await ad.load();
     }
   }
@@ -253,12 +324,14 @@ class EasyAds {
 
   Future _initFacebook({
     required bool iOSAdvertiserTrackingEnabled,
+    required bool testMode,
     String? testingId,
     String? interstitialPlacementId,
     String? rewardedPlacementId,
   }) async {
-    final status = await FacebookAudienceNetwork.init(
+    final status = await EasyAudienceNetwork.init(
         testingId: testingId,
+        testMode: testMode,
         iOSAdvertiserTrackingEnabled: iOSAdvertiserTrackingEnabled);
 
     _eventController.fireNetworkInitializedEvent(
@@ -288,37 +361,6 @@ class EasyAds {
     }
   }
 
-  /// Displays random ad network [adUnitType] ad.
-  /// It will randomly display one network and if that network's ad is not loaded, it will try second and so on until it exhaust all the network ads.
-  /// Returns bool indicating whether ad has been successfully displayed or not
-  ///
-  /// [adUnitType] should be mentioned here, only interstitial or rewarded should be mentioned here
-  bool showRandomAd(AdUnitType adUnitType) {
-    assert(
-        adUnitType == AdUnitType.interstitial ||
-            adUnitType == AdUnitType.rewarded,
-        'Only interstitial and rewarded types should be passed to this method');
-
-    final List<EasyAdBase> ads = (adUnitType == AdUnitType.rewarded
-            ? _rewardedAds
-            : _interstitialAds)
-        .toList(growable: false)
-      ..shuffle();
-
-    for (final ad in ads) {
-      if (ad.isAdLoaded) {
-        ad.show();
-        return true;
-      } else {
-        _logger.logInfo(
-            '${ad.adNetwork} ${ad.adUnitType} was not loaded, so called loading');
-        ad.load();
-      }
-    }
-
-    return false;
-  }
-
   /// Displays [adUnitType] ad from [adNetwork]. It will check if first ad it found from list is loaded,
   /// it will be displayed if [adNetwork] is not mentioned otherwise it will load the ad.
   ///
@@ -326,20 +368,52 @@ class EasyAds {
   ///
   /// [adUnitType] should be mentioned here, only interstitial or rewarded should be mentioned here
   /// if [adNetwork] is provided, only that network's ad would be displayed
-  /// if [random] is true, any random loaded ad would be displayed
-  bool showAd(AdUnitType adUnitType, {AdNetwork adNetwork = AdNetwork.any}) {
-    assert(
-        adUnitType == AdUnitType.interstitial ||
-            adUnitType == AdUnitType.rewarded,
-        'Only interstitial and rewarded types should be passed to this method');
+  /// if [shouldShowLoader] before interstitial. If it's true, you have to provide build context.
+  bool showAd(AdUnitType adUnitType,
+      {AdNetwork adNetwork = AdNetwork.any,
+      bool shouldShowLoader = false,
+      int delayInSeconds = 2,
+      BuildContext? context}) {
+    List<EasyAdBase> ads = [];
+    if (adUnitType == AdUnitType.rewarded) {
+      ads = _rewardedAds;
+    } else if (adUnitType == AdUnitType.interstitial) {
+      ads = _interstitialAds;
+    } else if (adUnitType == AdUnitType.appOpen) {
+      ads = _appOpenAds;
+    }
 
-    final List<EasyAdBase> ads =
-        adUnitType == AdUnitType.rewarded ? _rewardedAds : _interstitialAds;
+    if (adNetwork != AdNetwork.any) {
+      final ad = ads.firstWhereOrNull((e) => adNetwork == e.adNetwork);
+      if (ad?.isAdLoaded == true) {
+        if (ad?.adUnitType == AdUnitType.interstitial &&
+            shouldShowLoader &&
+            context != null) {
+          showLoaderDialog(context, delay: delayInSeconds)
+              .then((_) => ad?.show());
+        } else {
+          ad?.show();
+        }
+        return true;
+      } else {
+        _logger.logInfo(
+            '${ad?.adNetwork} ${ad?.adUnitType} was not loaded, so called loading');
+        ad?.load();
+        return false;
+      }
+    }
 
     for (final ad in ads) {
       if (ad.isAdLoaded) {
         if (adNetwork == AdNetwork.any || adNetwork == ad.adNetwork) {
-          ad.show();
+          if (ad.adUnitType == AdUnitType.interstitial &&
+              shouldShowLoader &&
+              context != null) {
+            showLoaderDialog(context, delay: delayInSeconds)
+                .then((_) => ad.show());
+          } else {
+            ad.show();
+          }
           return true;
         }
       } else {
@@ -356,17 +430,30 @@ class EasyAds {
   /// If a particular ad is already loaded, it will not load it again.
   /// Also you do not have to call this method everytime. Ad is automatically loaded after being displayed.
   ///
-  /// if [adNetwork] is provided, only that network's ad would be loaded
-  void loadAd({AdNetwork adNetwork = AdNetwork.any}) {
-    for (final e in _rewardedAds) {
-      if (adNetwork == AdNetwork.any || adNetwork == e.adNetwork) {
-        e.load();
+  /// if [adNetwork] is provided, only that network's ad will be loaded
+  /// if [adUnitType] is provided, only that unit type will be loaded, otherwise all unit types will be loaded
+  void loadAd({AdNetwork adNetwork = AdNetwork.any, AdUnitType? adUnitType}) {
+    if (adUnitType == null || adUnitType == AdUnitType.rewarded) {
+      for (final e in _rewardedAds) {
+        if (adNetwork == AdNetwork.any || adNetwork == e.adNetwork) {
+          e.load();
+        }
       }
     }
 
-    for (final e in _interstitialAds) {
-      if (adNetwork == AdNetwork.any || adNetwork == e.adNetwork) {
-        e.load();
+    if (adUnitType == null || adUnitType == AdUnitType.interstitial) {
+      for (final e in _interstitialAds) {
+        if (adNetwork == AdNetwork.any || adNetwork == e.adNetwork) {
+          e.load();
+        }
+      }
+    }
+
+    if (adUnitType == null || adUnitType == AdUnitType.appOpen) {
+      for (final e in _appOpenAds) {
+        if (adNetwork == AdNetwork.any || adNetwork == e.adNetwork) {
+          e.load();
+        }
       }
     }
   }
@@ -386,6 +473,16 @@ class EasyAds {
   /// if [adNetwork] is provided, only that network's ad would be checked
   bool isInterstitialAdLoaded({AdNetwork adNetwork = AdNetwork.any}) {
     final ad = _interstitialAds.firstWhereOrNull((e) =>
+        (adNetwork == AdNetwork.any || adNetwork == e.adNetwork) &&
+        e.isAdLoaded);
+    return ad?.isAdLoaded ?? false;
+  }
+
+  /// Returns bool indicating whether ad has been loaded
+  ///
+  /// if [adNetwork] is provided, only that network's ad would be checked
+  bool isAppOpenAdLoaded({AdNetwork adNetwork = AdNetwork.any}) {
+    final ad = _appOpenAds.firstWhereOrNull((e) =>
         (adNetwork == AdNetwork.any || adNetwork == e.adNetwork) &&
         e.isAdLoaded);
     return ad?.isAdLoaded ?? false;