+++ lib/src/utils/easy_banner_ad.dart
@@ -1,5 +1,6 @@
-import 'package:flutter/material.dart';
 import 'package:easy_ads_flutter/easy_ads_flutter.dart';
+import 'package:easy_ads_flutter/src/utils/badged_banner.dart';
+import 'package:flutter/material.dart';
 
 class EasyBannerAd extends StatefulWidget {
   final AdNetwork adNetwork;
@@ -8,31 +9,57 @@ class EasyBannerAd extends StatefulWidget {
     this.adNetwork = AdNetwork.admob,
     this.adSize = AdSize.banner,
     Key? key,
-  })  : assert(adNetwork != AdNetwork.appLovin,
-            "Applovin banner is not supported yet"),
-        super(key: key);
+  }) : super(key: key);
 
   @override
-  _EasyBannerAdState createState() => _EasyBannerAdState();
+  State<EasyBannerAd> createState() => _EasyBannerAdState();
 }
 
 class _EasyBannerAdState extends State<EasyBannerAd> {
-  late final EasyAdBase? _bannerAd = EasyAds.instance
-      .createBanner(adNetwork: widget.adNetwork, adSize: widget.adSize);
+  EasyAdBase? _bannerAd;
 
   @override
-  Widget build(BuildContext context) => _bannerAd?.show() ?? const SizedBox();
+  Widget build(BuildContext context) {
+    if (EasyAds.instance.showAdBadge) {
+      return BadgedBanner(child: _bannerAd?.show(), adSize: widget.adSize);
+    }
+
+    return _bannerAd?.show() ??
+        Container(height: widget.adSize.height.toDouble());
+  }
+
+  @override
+  void didUpdateWidget(covariant EasyBannerAd oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    createBanner();
+    _bannerAd?.onBannerAdReadyForSetState = onBannerAdReadyForSetState;
+  }
+
+  void createBanner() {
+    _bannerAd = EasyAds.instance
+        .createBanner(adNetwork: widget.adNetwork, adSize: widget.adSize);
+    _bannerAd?.load();
+  }
 
   @override
   void initState() {
-    _bannerAd?.load();
-
     super.initState();
+
+    createBanner();
+
+    _bannerAd?.onAdLoaded = onBannerAdReadyForSetState;
   }
 
   @override
   void dispose() {
     super.dispose();
     _bannerAd?.dispose();
+    _bannerAd = null;
+  }
+
+  void onBannerAdReadyForSetState(
+      AdNetwork adNetwork, AdUnitType adUnitType, Object? data) {
+    setState(() {});
   }
 }